slash commands

+++ src/commands/Image/whowouldwin.js
@@ -1,5 +1,5 @@
 // Dependencies
-const { MessageEmbed } = require('discord.js'),
+const { Embed } = require('../../utils'),
 	fetch = require('node-fetch'),
 	Command = require('../../structures/Command.js');
 
@@ -20,23 +20,25 @@ module.exports = class WhoWouldWin extends Command {
 	// Run command
 	async run(bot, message, settings) {
 		// Get user
-		const member = message.getMember();
+		const members = await message.getImage();
+		if (!members[1]) return message.channel.error('misc:INCORRECT_FORMAT', { EXAMPLE: settings.prefix.concat(message.translate('image/ship:USAGE')) }).then(m => m.delete({ timeout: 5000 }));
 
-		// send 'waiting' message
-		const msg = await message.channel.send(`${message.checkEmoji() ? bot.customEmojis['loading'] : ''} ${bot.translate(settings.Language, 'IMAGE/GENERATING_IMAGE')}`);
+		// send 'waiting' message to show bot has recieved message
+		const msg = await message.channel.send(message.translate('misc:GENERATING_IMAGE', {
+			EMOJI: message.checkEmoji() ? bot.customEmojis['loading'] : '' }));
 
 		// Try and convert image
 		try {
-			const json = await fetch(encodeURI(`https://nekobot.xyz/api/imagegen?type=whowouldwin&user1=${member[0].user.displayAvatarURL({ format: 'png', size: 512 })}&user2=${member[1].user.displayAvatarURL({ format: 'png', size: 512 })}`)).then(res => res.json());
+			const json = await fetch(encodeURI(`https://nekobot.xyz/api/imagegen?type=whowouldwin&user1=${members[0]}&user2=${members[1]}`)).then(res => res.json());
 
 			// send image in embed
-			const embed = new MessageEmbed()
+			const embed = new Embed(bot, message.guild)
 				.setImage(json.message);
 			message.channel.send(embed);
 		} catch(err) {
 			if (message.deletable) message.delete();
 			bot.logger.error(`Command: '${this.help.name}' has error: ${err.message}.`);
-			message.channel.error(settings.Language, 'ERROR_MESSAGE', err.message).then(m => m.delete({ timeout: 5000 }));
+			message.channel.error('misc:ERROR_MESSAGE', { ERROR: err.message }).then(m => m.delete({ timeout: 5000 }));
 		}
 		msg.delete();
 	}
